fixed default expression traits problem

+++ cml/mathlib/misc.h
@@ -43,9 +43,9 @@ namespace cml {
 
 /** Return an N-d zero vector */
 template < size_t N >
-vector< int, fixed<N> > zero()
+vector< double, fixed<N> > zero()
 {
-    typedef vector< int, fixed<N> > vector_type;
+    typedef vector< double, fixed<N> > vector_type;
 
     vector_type result;
     result.zero();
@@ -54,12 +54,12 @@ vector< int, fixed<N> > zero()
 
 /** Return an N-d cardinal axis by index */
 template < size_t N >
-vector< int, fixed<N> > axis(size_t i)
+vector< double, fixed<N> > axis(size_t i)
 {
     /* Checking */
     detail::CheckValidArg(i < N);
 
-    typedef vector< int, fixed<N> > vector_type;
+    typedef vector< double, fixed<N> > vector_type;
     vector_type result;
     result.cardinal(i);
     return result;
@@ -67,9 +67,9 @@ vector< int, fixed<N> > axis(size_t i)
 
 /** Return an NxM zero matrix */
 template < size_t N, size_t M >
-matrix< int, fixed<N,M>, row_basis, row_major > zero()
+matrix< double, fixed<N,M>, row_basis, row_major > zero()
 {
-    typedef matrix< int, fixed<N,M>, row_basis, row_major > matrix_type;
+    typedef matrix< double, fixed<N,M>, row_basis, row_major > matrix_type;
 
     matrix_type result;
     result.zero();
@@ -78,9 +78,9 @@ matrix< int, fixed<N,M>, row_basis, row_major > zero()
 
 /** Return an NxN identity matrix */
 template < size_t N >
-matrix< int, fixed<N,N>, row_basis, row_major > identity()
+matrix< double, fixed<N,N>, row_basis, row_major > identity()
 {
-    typedef matrix< int, fixed<N,N>, row_basis, row_major > matrix_type;
+    typedef matrix< double, fixed<N,N>, row_basis, row_major > matrix_type;
 
     matrix_type result;
     result.identity();
@@ -89,9 +89,9 @@ matrix< int, fixed<N,N>, row_basis, row_major > identity()
 
 /** Return an NxM identity transform */
 template < size_t N, size_t M >
-matrix< int, fixed<N,M>, row_basis, row_major > identity_transform()
+matrix< double, fixed<N,M>, row_basis, row_major > identity_transform()
 {
-    typedef matrix< int, fixed<N,M>, row_basis, row_major > matrix_type;
+    typedef matrix< double, fixed<N,M>, row_basis, row_major > matrix_type;
 
     matrix_type result;
     identity_transform(result);
@@ -103,17 +103,17 @@ matrix< int, fixed<N,M>, row_basis, row_major > identity_transform()
 //////////////////////////////////////////////////////////////////////////////
 
 /** Return the 2D zero vector */
-inline vector< int, fixed<2> > zero_2D() {
+inline vector< double, fixed<2> > zero_2D() {
     return zero<2>();
 }
 
 /** Return the 3D zero vector */
-inline vector< int, fixed<3> > zero_3D() {
+inline vector< double, fixed<3> > zero_3D() {
     return zero<3>();
 }
 
 /** Return the 4D zero vector */
-inline vector< int, fixed<4> > zero_4D() {
+inline vector< double, fixed<4> > zero_4D() {
     return zero<4>();
 }
 
@@ -122,37 +122,37 @@ inline vector< int, fixed<4> > zero_4D() {
 //////////////////////////////////////////////////////////////////////////////
 
 /** Return a 2D cardinal axis by index */
-inline vector< int, fixed<2> > axis_2D(size_t i) {
+inline vector< double, fixed<2> > axis_2D(size_t i) {
     return axis<2>(i);
 }
 
 /** Return a 3D cardinal axis by index */
-inline vector< int, fixed<3> > axis_3D(size_t i) {
+inline vector< double, fixed<3> > axis_3D(size_t i) {
     return axis<3>(i);
 }
 
 /** Return a the 2D x cardinal axis */
-inline vector< int, fixed<2> > x_axis_2D() {
+inline vector< double, fixed<2> > x_axis_2D() {
     return axis_2D(0);
 }
 
 /** Return a the 2D y cardinal axis */
-inline vector< int, fixed<2> > y_axis_2D() {
+inline vector< double, fixed<2> > y_axis_2D() {
     return axis_2D(1);
 }
 
 /** Return a the 3D x cardinal axis */
-inline vector< int, fixed<3> > x_axis_3D() {
+inline vector< double, fixed<3> > x_axis_3D() {
     return axis_3D(0);
 }
 
 /** Return a the 3D y cardinal axis */
-inline vector< int, fixed<3> > y_axis_3D() {
+inline vector< double, fixed<3> > y_axis_3D() {
     return axis_3D(1);
 }
 
 /** Return a the 3D z cardinal axis */
-inline vector< int, fixed<3> > z_axis_3D() {
+inline vector< double, fixed<3> > z_axis_3D() {
     return axis_3D(2);
 }
 
@@ -161,17 +161,17 @@ inline vector< int, fixed<3> > z_axis_3D() {
 //////////////////////////////////////////////////////////////////////////////
 
 /** Return the 2x2 zero matrix */
-inline matrix< int, fixed<2,2>, row_basis, row_major > zero_2x2() {
+inline matrix< double, fixed<2,2>, row_basis, row_major > zero_2x2() {
     return zero<2,2>();
 }
 
 /** Return the 3x3 zero matrix */
-inline matrix< int, fixed<3,3>, row_basis, row_major > zero_3x3() {
+inline matrix< double, fixed<3,3>, row_basis, row_major > zero_3x3() {
     return zero<3,3>();
 }
 
 /** Return the 4x4 zero matrix */
-inline matrix< int, fixed<4,4>, row_basis, row_major > zero_4x4() {
+inline matrix< double, fixed<4,4>, row_basis, row_major > zero_4x4() {
     return zero<4,4>();
 }
 
@@ -180,17 +180,17 @@ inline matrix< int, fixed<4,4>, row_basis, row_major > zero_4x4() {
 //////////////////////////////////////////////////////////////////////////////
 
 /** Return the 2x2 identity matrix */
-inline matrix< int, fixed<2,2>, row_basis, row_major > identity_2x2() {
+inline matrix< double, fixed<2,2>, row_basis, row_major > identity_2x2() {
     return identity<2>();
 }
 
 /** Return the 3x3 identity matrix */
-inline matrix< int, fixed<3,3>, row_basis, row_major > identity_3x3() {
+inline matrix< double, fixed<3,3>, row_basis, row_major > identity_3x3() {
     return identity<3>();
 }
 
 /** Return the 4x4 identity matrix */
-inline matrix< int, fixed<4,4>, row_basis, row_major > identity_4x4() {
+inline matrix< double, fixed<4,4>, row_basis, row_major > identity_4x4() {
     return identity<4>();
 }
 
@@ -199,22 +199,22 @@ inline matrix< int, fixed<4,4>, row_basis, row_major > identity_4x4() {
 //////////////////////////////////////////////////////////////////////////////
 
 /** Return a 3x2 identity transform */
-inline matrix< int,fixed<3,2>,row_basis,row_major > identity_transform_3x2() {
+inline matrix< double,fixed<3,2>,row_basis,row_major > identity_transform_3x2() {
     return identity_transform<3,2>();
 }
 
 /** Return a 2x3 identity transform */
-inline matrix< int,fixed<2,3>,col_basis,col_major > identity_transform_2x3() {
+inline matrix< double,fixed<2,3>,col_basis,col_major > identity_transform_2x3() {
     return identity_transform<2,3>();
 }
 
 /** Return a 4x3 identity transform */
-inline matrix< int,fixed<4,3>,row_basis,row_major > identity_transform_4x3() {
+inline matrix< double,fixed<4,3>,row_basis,row_major > identity_transform_4x3() {
     return identity_transform<4,3>();
 }
 
 /** Return a 3x4 identity transform */
-inline matrix< int,fixed<3,4>,col_basis,col_major > identity_transform_3x4() {
+inline matrix< double,fixed<3,4>,col_basis,col_major > identity_transform_3x4() {
     return identity_transform<3,4>();
 }
